* commit 'b92eab4f34fafedda96d924a93d2ffce3b6d4dcb':
  Get rid of some compiler warnings, mainly from Visual C++

GitOrigin-RevId: fdbfcde2c3bc690bbcea580b058688138ea2d7ea

+++ test/core/objects/ActionCommonImplTest.cxx
@@ -790,7 +790,7 @@ TEST_F(ActionCommonImplTest, leavingAnActionSetsTheEndTime)
 TEST_F(ActionCommonImplTest, leavingAnActionSetsTheEndSequenceNumber)
 {
 	// with
-	int64_t endSequenceNumber = 1234;
+	int32_t endSequenceNumber = 1234;
 
 	// expect
 	EXPECT_CALL(*mockNiceBeacon, createSequenceNumber())
